# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suxen control-plane API",
    "version": "dev",
    "description": "Administrative API for repositories, access control, retention, provenance, and operations. OCI Distribution endpoints under /v2 and /repository/{name}/v2 use the OCI error envelope and are intentionally outside this control-plane contract."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current suxen server"
    }
  ],
  "security": [
    {},
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    },
    {
      "oidcSession": []
    }
  ],
...
  "paths": {
...
    "/api/v1/webhooks/{name}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WebhookName"
        }
      ],
      "get": {
        "operationId": "getWebhook",
        "x-suxen-required-privilege": "admin:webhooks:read",
        "summary": "Get a webhook",
        "tags": [
          "Webhooks"
        ],
        "responses": {
          "200": {
            "description": "Webhook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Webhook"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "putWebhook",
        "x-suxen-required-privilege": "admin:webhooks:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Create or replace a webhook by name",
        "description": "The path name is authoritative. Omitting secret preserves the existing secret; a new webhook requires one. Set enabled to false to deactivate delivery without deleting the subscription.",
        "tags": [
          "Webhooks"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/WebhookUpdate"
        },
        "responses": {
          "200": {
            "description": "Webhook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Webhook"
                }
              }
            }
          },
          "201": {
            "description": "Webhook created",
            "headers": {"Location": {"$ref": "#/components/headers/Location"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Webhook"}}}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteWebhook",
        "x-suxen-required-privilege": "admin:webhooks:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Delete a webhook",
        "tags": [
          "Webhooks"
        ],
        "responses": {
          "204": {
            "description": "Webhook deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/verify": {
      "post": {
        "operationId": "verifyBlobStores",
        "x-suxen-required-privilege": "admin:verify:run",
        "summary": "Verify blob store integrity",
        "description": "Cross-checks each blob store's contents against the digests the metadata references and reports dangling references, orphans, and (when re-hashing) content that no longer matches its digest. Read-only; mutates nothing. Task history remains under /api/v1/tasks.",
        "tags": [
          "Operations"
        ],
        "parameters": [
          {
            "name": "blobStore",
            "in": "query",
            "required": false,
            "description": "Verify only this blob store; omit to verify every store.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rehash",
            "in": "query",
            "required": false,
            "description": "Re-hash present, referenced blobs to detect bit-rot.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blob store verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobStoreVerifyResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Local username and password."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token or OIDC JWT",
        "description": "Suxen API token or a configured OIDC bearer token. API-token scopes use repository:<name|*>:<read|write|delete|annotate|manage> and admin:<resource|*>:<action|*>. An empty scope list does not further restrict the user's roles.",
        "x-suxen-scope-syntax": [
          "repository:<name|*>:<read|write|delete|annotate|manage>",
          "admin:<resource|*>:<action|*>"
        ]
      },
      "oidcSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suxen_oidc_token",
        "description": "HTTP-only browser session established by the configured OIDC login flow."
      }
    },
    "headers": {
      "Location": {
        "description": "Canonical path of the created resource.",
        "schema": {"type": "string"}
      }
    },
    "parameters": {
...
      "WebhookName": {
        "name": "name",
        "in": "path",
        "required": true,
        "description": "Webhook provisioning name",
        "schema": {"$ref": "#/components/schemas/ResourceName"}
      },
...
      "Force": {
        "name": "force",
        "in": "query",
        "required": false,
        "description": "Transfer a managed resource out of declarative provisioning before applying the mutation",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
    "requestBodies": {
...
      "WebhookUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/NamedWebhookRequest"}
          }
        }
      }
    },
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request. Account validation codes include invalid_username and password_required; the reserved repository name default returns reserved_repository_name.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required. Control-plane routes answer 401 without a WWW-Authenticate challenge so browsers do not raise the native Basic-auth popup; clients authenticate with a bearer token or the cookie session and treat 401 as a prompt to sign in.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient privilege",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists, is declaratively managed, or is still referenced",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "ResourceName": {
        "type": "string",
        "maxLength": 255,
        "pattern": "^[a-z0-9_-]+$",
        "example": "releases"
      },
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 7807 problem details with a stable suxen error code.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:suxen:problem:not_found"
          },
          "title": {
            "type": "string",
            "example": "Not found"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9_]+$",
            "example": "not_found"
          }
        }
      },
...
      "NamedWebhookRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["url", "events"],
        "properties": {
          "name": {
            "allOf": [{"$ref": "#/components/schemas/ResourceName"}],
            "description": "Ignored when the webhook name is supplied in the path."
          },
          "url": {"type": "string", "format": "uri"},
          "secret": {"type": "string", "minLength": 16, "writeOnly": true},
          "events": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "asset.uploaded",
                "asset.deleted",
                "asset.downloaded",
                "component.created",
                "cleanup.completed"
              ]
            }
          },
          "repositories": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ResourceName"}
          },
          "enabled": {"type": "boolean", "default": true}
        }
      },
...
      "Webhook": {
        "type": "object",
        "required": [
          "name",
          "url",
          "events",
          "enabled",
          "managed",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ResourceName"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "repositories": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ResourceName"}
          },
          "enabled": {
            "type": "boolean"
          },
          "managed": {
            "type": "boolean",
            "readOnly": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "VerifyFinding": {
        "type": "object",
        "required": [
          "blobStore",
          "digest"
        ],
        "properties": {
          "blobStore": {
            "type": "string"
          },
          "digest": {
            "type": "string"
          }
        }
      },
      "BlobStoreVerifyResult": {
        "type": "object",
        "required": [
          "rehash",
          "stores",
          "scanned",
          "referenced",
          "rehashed",
          "danglingTotal",
          "orphanedTotal",
          "mismatchedTotal"
        ],
        "properties": {
          "rehash": {
            "type": "boolean"
          },
          "blobStore": {
            "type": "string"
          },
          "stores": {
            "type": "integer",
            "minimum": 0
          },
          "scanned": {
            "type": "integer",
            "minimum": 0
          },
          "referenced": {
            "type": "integer",
            "minimum": 0
          },
          "rehashed": {
            "type": "integer",
            "minimum": 0
          },
          "danglingTotal": {
            "type": "integer",
            "minimum": 0
          },
          "orphanedTotal": {
            "type": "integer",
            "minimum": 0
          },
          "mismatchedTotal": {
            "type": "integer",
            "minimum": 0
          },
          "dangling": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerifyFinding"
            }
          },
          "mismatched": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerifyFinding"
            }
          }
        }
      }
    }
  }
}